move all object definitions to lunatik_obj.h
